rename: 파일 혹은 폴더명을 수정 (#57)
- .gitkeep 파일들 전부 삭제
- 파일명, 폴더명 변경 및 이동
